pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chonkie[all]>=1.1.1",
    "dataframe-image>=0.2.7",
    "dotenv>=0.9.9",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-huggingface>=0.3.1",
    "lxml>=6.0.0",
    "matplotlib>=3.10.5",
    "neo4j>=5.28.2",
    "numpy>=2.3.2",
    "openai>=1.97.1",
    "outlines>=1.1.1",
    "pandas>=2.3.1",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pydantic-xml>=2.17.3",
    "pytest>=8.4.1",
    "sentence-transformers>=5.0.0",
    "tiktoken>=0.9.0",
    "torch>=2.7.1",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]

[dependency-groups]
dev = [
    "ruff>=0.12.7",
]


[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.ruff]
line-length = 2000
